Add 5G Radio LMS#2 CFR equalizer controls

+++ src/boards/IComms.h
@@ -5,7 +5,7 @@ class IComms
 {
   public:
     IComms(){};
-    virtual void SPI(uint32_t spiBusAddress, const uint32_t *MOSI, uint32_t *MISO, size_t count) = 0;
+    virtual void SPI(uint32_t spiBusAddress, const uint32_t *MOSI, uint32_t *MISO, uint32_t count) = 0;
 
     /*!
      * Write to an available I2C slave.
@@ -14,7 +14,7 @@ class IComms
      * @param length output data length
      * @return 0-success
      */
-    virtual int I2CWrite(int address, const uint8_t *data, size_t length) = 0;
+    virtual int I2CWrite(int address, const uint8_t *data, uint32_t length) = 0;
 
     /*!
      * Read from an available I2C slave.
@@ -27,7 +27,7 @@ class IComms
      * \param length Number of bytes to read
      * @return 0-success
      */
-    virtual int I2CRead(int addres, uint8_t *dest, size_t length) = 0;
+    virtual int I2CRead(int addres, uint8_t *dest, uint32_t length) = 0;
 };
 
 } // namespace lime
